Resolved next commit

+++ src/adapters/pouch.http.js
@@ -172,6 +172,9 @@ var HttpPouch = function(opts, callback) {
     });
   };
   if (!opts.skipSetup) {
+    // Skipping setup so force db to be ready
+    //api.taskqueue.ready(true);
+    //api.taskqueue.execute(api);
     ajax({auth: host.auth, method: 'GET', url: db_url}, function(err, ret) {
       //check if the db exists
       if (err) {
@@ -198,6 +201,10 @@ var HttpPouch = function(opts, callback) {
   };
 
   api.request = function(options, callback) {
+    if (!api.taskqueue.ready()) {
+      api.taskqueue.addJob('request', arguments);
+      return;
+    }
     options.auth = host.auth;
     options.url = genDBUrl(host, options.url);
     ajax(options, callback);
@@ -206,6 +213,10 @@ var HttpPouch = function(opts, callback) {
   // Sends a POST request to the host calling the couchdb _compact function
   //    version: The version of CouchDB it is running
   api.compact = function(callback) {
+    if (!api.taskqueue.ready()) {
+      api.taskqueue.addJob('compact', arguments);
+      return;
+    }
     ajax({
       auth: host.auth,
       url: genDBUrl(host, '_compact'),
@@ -217,6 +228,10 @@ var HttpPouch = function(opts, callback) {
   //    couchdb: A welcome string
   //    version: The version of CouchDB it is running
   api.info = function(callback) {
+    if (!api.taskqueue.ready()) {
+      api.taskqueue.addJob('info', arguments);
+      return;
+    }
     ajax({
       auth: host.auth,
       method:'GET',
@@ -229,8 +244,8 @@ var HttpPouch = function(opts, callback) {
   // _design/ID or _local/ID path
   api.get = function(id, opts, callback) {
 
-    if (!Pouch.taskqueue.ready) {
-      Pouch.taskqueue.addJob('get', arguments);
+    if (!api.taskqueue.ready()) {
+      api.taskqueue.addJob('get', arguments);
       return;
     }
     // If no options were given, set the callback to the second parameter
@@ -327,6 +342,10 @@ var HttpPouch = function(opts, callback) {
 
   // Delete the document given by doc from the database given by host.
   api.remove = function(doc, opts, callback) {
+    if (!api.taskqueue.ready()) {
+      api.taskqueue.addJob('remove', arguments);
+      return;
+    }
     // If no options were given, set the callback to be the second parameter
     if (typeof opts === 'function') {
       callback = opts;
@@ -343,6 +362,10 @@ var HttpPouch = function(opts, callback) {
 
   // Remove the attachment given by the id and rev
   api.removeAttachment = function idb_removeAttachment(id, rev, callback) {
+    if (!api.taskqueue.ready()) {
+      api.taskqueue.addJob('removeAttachment', arguments);
+      return;
+    }
     ajax({
       auth: host.auth,
       method: 'DELETE',
@@ -354,6 +377,10 @@ var HttpPouch = function(opts, callback) {
   // to the document with the given id, the revision given by rev, and
   // add it to the database given by host.
   api.putAttachment = function(id, rev, blob, type, callback) {
+    if (!api.taskqueue.ready()) {
+      api.taskqueue.addJob('putAttachment', arguments);
+      return;
+    }
     if (typeof type === 'function') {
       callback = type;
       type = blob;
@@ -368,7 +395,7 @@ var HttpPouch = function(opts, callback) {
     var url = genDBUrl(host, id);
     if (rev) {
       url += '?rev=' + rev;
-    }
+
     // Add the attachment
     ajax({
       auth: host.auth,
@@ -383,6 +410,10 @@ var HttpPouch = function(opts, callback) {
   // Add the document given by doc (in JSON string format) to the database
   // given by host. This fails if the doc has no _id field.
   api.put = function(doc, opts, callback) {
+    if (!api.taskqueue.ready()) {
+      api.taskqueue.addJob('put', arguments);
+      return;
+    }
     // If no options were given, set the callback to be the second parameter
     if (typeof opts === 'function') {
       callback = opts;
@@ -422,6 +453,10 @@ var HttpPouch = function(opts, callback) {
   // given by host. This does not assume that doc is a new document (i.e. does not
   // have a _id or a _rev field.
   api.post = function(doc, opts, callback) {
+    if (!api.taskqueue.ready()) {
+      api.taskqueue.addJob('post', arguments);
+      return;
+    }
     // If no options were given, set the callback to be the second parameter
     if (typeof opts === 'function') {
       callback = opts;
@@ -449,8 +484,8 @@ var HttpPouch = function(opts, callback) {
   // given by host.
   api.bulkDocs = function(req, opts, callback) {
 
-    if (!Pouch.taskqueue.ready) {
-      Pouch.taskqueue.addJob('bulkDocs', arguments);
+    if (!api.taskqueue.ready()) {
+      api.taskqueue.addJob('bulkDocs', arguments);
       return;
     }
     // If no options were given, set the callback to be the second parameter
@@ -484,8 +519,8 @@ var HttpPouch = function(opts, callback) {
   // by host and ordered by increasing id.
   api.allDocs = function(opts, callback) {
     // If no options were given, set the callback to be the second parameter
-    if (!Pouch.taskqueue.ready) {
-      Pouch.taskqueue.addJob('allDocs', arguments);
+    if (!api.taskqueue.ready()) {
+      api.taskqueue.addJob('allDocs', arguments);
       return;
     }
     if (typeof opts === 'function') {
@@ -558,6 +593,10 @@ var HttpPouch = function(opts, callback) {
   // TODO According to the README, there should be two other methods here,
   // api.changes.addListener and api.changes.removeListener.
   api.changes = function(opts) {
+    if (!api.taskqueue.ready()) {
+      api.taskqueue.addJob('changes', arguments);
+      return;
+    }
 
     if (Pouch.DEBUG) {
       console.log(db_url + ': Start Changes Feed: continuous=' + opts.continuous);
@@ -716,6 +755,10 @@ var HttpPouch = function(opts, callback) {
   // those that do NOT correspond to revisions stored in the database.
   // See http://wiki.apache.org/couchdb/HttpPostRevsDiff
   api.revsDiff = function(req, opts, callback) {
+    if (!api.taskqueue.ready()) {
+      api.taskqueue.addJob('revsDiff', arguments);
+      return;
+    }
     // If no options were given, set the callback to be the second parameter
     if (typeof opts === 'function') {
       callback = opts;
@@ -734,6 +777,10 @@ var HttpPouch = function(opts, callback) {
   };
 
   api.close = function(callback) {
+    if (!api.taskqueue.ready()) {
+      api.taskqueue.addJob('close', arguments);
+      return;
+    }
     call(callback, null);
   };
